# Implement io.ReaderFrom/WriterTo for Conn