@pires 2021-02-20
Thank you, Fabian. Can you, please, implement example tests so to have code others can learn from while keeping up with current code coverage?