@pires 2021-02-20
Maybe even benchmark tests to measure the actual performance gains?